+++ lib/source_gen.dart
@@ -8,6 +8,7 @@ export 'src/constants/revive.dart' show Revivable;
 export 'src/generator.dart';
 export 'src/generator_for_annotation.dart';
 export 'src/library.dart' show AnnotatedElement, LibraryReader;
+export 'src/sorted_fields.dart' show sortedFields;
 export 'src/span_for_element.dart' show spanForElement;
 export 'src/type_checker.dart' show TypeChecker, UnresolvedAnnotationException;
 export 'src/utils.dart' show typeNameOf;