+++ lib/source_gen.dart
@@ -5,6 +5,7 @@
 export 'src/builder.dart';
 export 'src/constants/reader.dart' show ConstantReader;
 export 'src/constants/revive.dart' show Revivable;
+export 'src/dart_string_literal.dart';
 export 'src/generator.dart';
 export 'src/generator_for_annotation.dart';
 export 'src/library.dart' show AnnotatedElement, LibraryReader;